requirements.txt
# ref: https://github.com/Rapptz/discord.py/issues/1249
# pip install -U git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]
# discord.py==discord.py==1.0.0a1792+gdea3ba5
PyYAML==5.1
Click==7.0
google-api-python-client==1.7.8
pandas==0.24.2